Rename move and fix tests
